refactoring last updated function
