+++ src/data/constants.js
@@ -1,0 +1,1 @@
+export const TIME_FORMAT = "MMMM d yyyy, HH:mm";
